requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
google-cloud-speech==2.21.0
sqlalchemy==2.0.23
streamlit==1.28.0
audio-recorder-streamlit==0.0.8
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pydub==0.25.1

